+++ lib/src/fastlog_base.dart
@@ -1,3 +1,4 @@
+import 'dart:async';
 import 'dart:convert';
 import 'generate_log.dart';
 
@@ -7,12 +8,13 @@ class FastLog {
   static bool _isColored = true;
   static bool _showTime = false;
   static bool _useEmoji = true;
-  static String _logLevel = "INFO";
+  static String _logLevel = "TRACE";
   static int _messageLimit = 200; // Default character limit
 
+  static final StreamController<String> _logController =
+      StreamController.broadcast();
+
   /// Log levels in priority order.
-  ///
-  /// The log levels are ordered by priority: TRACE < DEBUG < INFO < WARN < ERROR < FATAL.
   static const List<String> _levels = [
     "TRACE",
     "DEBUG",
@@ -22,27 +24,15 @@ class FastLog {
     "FATAL"
   ];
 
-  /// Default emoji indicators for each log level.
-  ///
-  /// These emojis are used for visual indication of log levels when `useEmoji` is enabled.
   static final Map<String, String> _emojis = {
-    "TRACE": "🕵️", // Spyglass for tracing
-    "DEBUG": "🔧", // Wrench for debugging
-    "INFO": "📢", // Loudspeaker for general information
-    "WARN": "⚠️", // Warning sign for warnings
-    "ERROR": "🚨", // Alarm for errors
-    "FATAL": "☠️", // Skull for fatal errors
+    "TRACE": "🕵️",
+    "DEBUG": "🔧",
+    "INFO": "📢",
+    "WARN": "⚠️",
+    "ERROR": "🚨",
+    "FATAL": "☠️",
   };
 
-  /// Configure global logging settings.
-  ///
-  /// - `justDebug`: If true, only debug logs are shown. Default is true.
-  /// - `isColored`: If true, log output will be colorized. Default is true.
-  /// - `showTime`: If true, log timestamps will be displayed. Default is false.
-  /// - `useEmoji`: If true, emojis will be included in log output. Default is true.
-  /// - `logLevel`: The default log level to use. Logs below this level will be ignored.
-  /// - `messageLimit`: The maximum number of characters for log messages.
-  /// - `customEmojis`: A map of custom emojis to override the default ones for each log level.
   static void config({
     bool? justDebug,
     bool? isColored,
@@ -63,12 +53,6 @@ class FastLog {
     if (customEmojis != null) _emojis.addAll(customEmojis);
   }
 
-  /// Validate and set the log level.
-  ///
-  /// This method ensures that the provided log level is valid and exists in the `_levels` list.
-  /// If the level is valid, it is set as the active log level.
-  ///
-  /// - `level`: The log level to set (e.g., "INFO", "DEBUG").
   static void setLogLevel(String level) {
     final upperLevel = level.toUpperCase();
     if (_levels.contains(upperLevel)) {
@@ -76,14 +60,6 @@ class FastLog {
     }
   }
 
-  /// Safely convert message to a string.
-  ///
-  /// This method ensures that any message is properly converted into a string representation.
-  /// If the message is a non-serializable object, it will be converted into a readable format.
-  ///
-  /// - `message`: The message to convert.
-  ///
-  /// Returns a string representation of the message.
   static String _convertMessage(dynamic message) {
     try {
       if (message is String) return message;
@@ -93,14 +69,6 @@ class FastLog {
     }
   }
 
-  /// Helper function to apply message limit.
-  ///
-  /// This method ensures that the log message does not exceed the predefined character limit.
-  /// If the message exceeds the limit, it will be truncated and an ellipsis will be appended.
-  ///
-  /// - `message`: The message to apply the limit to.
-  ///
-  /// Returns the formatted message.
   static String _limitMessage(dynamic message) {
     final String convertedMessage = _convertMessage(message);
     return convertedMessage.length > _messageLimit
@@ -108,61 +76,31 @@ class FastLog {
         : convertedMessage;
   }
 
-  /// Generate log messages for the TRACE log level.
-  ///
-  /// - `message`: The message to log.
-  ///
-  /// Returns the generated log message for the TRACE level.
-  static String trace(dynamic message) => _log("TRACE", message);
+  static String _limitTag(String tag, int limitValue) {
+    final String convertedMessage = _convertMessage(tag);
+    return convertedMessage.length > limitValue
+        ? "${convertedMessage.substring(0, limitValue)}.."
+        : convertedMessage;
+  }
 
-  /// Generate log messages for the DEBUG log level.
-  ///
-  /// - `message`: The message to log.
-  ///
-  /// Returns the generated log message for the DEBUG level.
-  static String debug(dynamic message) => _log("DEBUG", message);
+  static Future<void> trace(dynamic message, {String tag = ""}) async =>
+      await _log("TRACE", message, tag);
+  static Future<void> debug(dynamic message, {String tag = ""}) async =>
+      await _log("DEBUG", message, tag);
+  static Future<void> info(dynamic message, {String tag = ""}) async =>
+      await _log("INFO", message, tag);
+  static Future<void> warn(dynamic message, {String tag = ""}) async =>
+      await _log("WARN", message, tag);
+  static Future<void> error(dynamic message, {String tag = ""}) async =>
+      await _log("ERROR", message, tag);
+  static Future<void> fatal(dynamic message, {String tag = ""}) async =>
+      await _log("FATAL", message, tag);
 
-  /// Generate log messages for the INFO log level.
-  ///
-  /// - `message`: The message to log.
-  ///
-  /// Returns the generated log message for the INFO level.
-  static String info(dynamic message) => _log("INFO", message);
-
-  /// Generate log messages for the WARN log level.
-  ///
-  /// - `message`: The message to log.
-  ///
-  /// Returns the generated log message for the WARN level.
-  static String warn(dynamic message) => _log("WARN", message);
-
-  /// Generate log messages for the ERROR log level.
-  ///
-  /// - `message`: The message to log.
-  ///
-  /// Returns the generated log message for the ERROR level.
-  static String error(dynamic message) => _log("ERROR", message);
-
-  /// Generate log messages for the FATAL log level.
-  ///
-  /// - `message`: The message to log.
-  ///
-  /// Returns the generated log message for the FATAL level.
-  static String fatal(dynamic message) => _log("FATAL", message);
-
-  /// Internal log function to generate log messages for a given level.
-  ///
-  /// This function assembles the log message based on the provided level and message.
-  /// It incorporates the global settings like color, emoji, timestamp, and log level.
-  ///
-  /// - `level`: The log level to generate (e.g., "INFO", "DEBUG").
-  /// - `message`: The message to log.
-  ///
-  /// Returns the formatted log message.
-  static String _log(String level, dynamic message) {
-    return generateLog(
+  static Future<void> _log(String level, dynamic message, String tag) async {
+    final logMessage = generateLog(
       _limitMessage(message),
       level,
+      _limitTag(tag, 7),
       _justDebug,
       _showTime,
       _useEmoji,
@@ -171,5 +109,10 @@ class FastLog {
       _levels,
       _emojis,
     );
+
+    _logController.add(logMessage);
+    await Future.delayed(Duration.zero, () => print(logMessage));
   }
+
+  static Stream<String> get logStream => _logController.stream;
 }