+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 ## 0.1.3
 
-- Initial release of `fastlog`, a super lightweight and high-performance logging package for Dart and Flutter.
+- Initial release of `flutter_fastlog`, a super lightweight and high-performance logging package for Dart and Flutter.